Code cleanup
