+++ src/net/erbros/Lottery/PlayerJoinListener.java
@@ -6,13 +6,14 @@ import org.bukkit.event.player.PlayerListener;
 import net.erbros.Lottery.Lottery;
 
 public class PlayerJoinListener extends PlayerListener {
-	public PlayerJoinListener() { }
-		
+	public PlayerJoinListener() {
+	}
+
 	public void onPlayerJoin(PlayerJoinEvent event) {
 		// Send the player some info about time until lottery draw?
-		event.getPlayer().sendMessage(ChatColor.GOLD + "[LOTTERY] " + ChatColor.WHITE + "Draw in: " + ChatColor.RED + Lottery.timeUntil(Lottery.nextexec));
+		event.getPlayer().sendMessage(
+				ChatColor.GOLD + "[LOTTERY] " + ChatColor.WHITE + "Draw in: "
+						+ ChatColor.RED + Lottery.timeUntil(Lottery.nextexec));
 	}
-	
+
 }
-	
-
