+++ src/net/erbros/Lottery/PluginListener.java
@@ -6,37 +6,41 @@ import org.bukkit.event.server.PluginDisableEvent;
 import org.bukkit.event.server.PluginEnableEvent;
 import org.bukkit.event.server.ServerListener;
 
-
 public class PluginListener extends ServerListener {
-    private Lottery plugin;
-    private Methods Methods = null;
+	private Lottery plugin;
+	private Methods Methods = null;
 
-    public PluginListener(Lottery plugin) {
-        this.plugin = plugin;
-        this.Methods = new Methods();
-    }
+	public PluginListener(Lottery plugin) {
+		this.plugin = plugin;
+		this.Methods = new Methods();
+	}
 
-    @Override
-    public void onPluginDisable(PluginDisableEvent event) {
-        if (this.Methods != null && this.Methods.hasMethod()) {
-            Boolean check = this.Methods.checkDisabled(event.getPlugin());
+	@Override
+	public void onPluginDisable(PluginDisableEvent event) {
+		if (this.Methods != null && this.Methods.hasMethod()) {
+			Boolean check = this.Methods.checkDisabled(event.getPlugin());
 
-            if(check) {
-                this.plugin.Method = null;
-                System.out.println("[Lottery] Payment method was disabled. No longer accepting payments.");
-            }
-        }
-    }
+			if (check) {
+				this.plugin.Method = null;
+				System.out
+						.println("[Lottery] Payment method was disabled. No longer accepting payments.");
+			}
+		}
+	}
 
-    @Override
-    public void onPluginEnable(PluginEnableEvent event) {
-        if (!this.Methods.hasMethod()) {
-            if(this.Methods.setMethod(event.getPlugin())) {
-                // You might want to make this a public variable inside your MAIN class public Method Method = null;
-                // then reference it through this.plugin.Method so that way you can use it in the rest of your plugin ;)
-                this.plugin.Method = this.Methods.getMethod();
-                System.out.println("[Lottery] Payment method found (" + this.plugin.Method.getName() + " version: " + this.plugin.Method.getVersion() + ")");
-            }
-        }
-    }
+	@Override
+	public void onPluginEnable(PluginEnableEvent event) {
+		if (!this.Methods.hasMethod()) {
+			if (this.Methods.setMethod(event.getPlugin())) {
+				// You might want to make this a public variable inside your
+				// MAIN class public Method Method = null;
+				// then reference it through this.plugin.Method so that way you
+				// can use it in the rest of your plugin ;)
+				this.plugin.Method = this.Methods.getMethod();
+				System.out.println("[Lottery] Payment method found ("
+						+ this.plugin.Method.getName() + " version: "
+						+ this.plugin.Method.getVersion() + ")");
+			}
+		}
+	}
 }
